Task4

+++ Tyuiu.ZhdanovaAA.Sprint5.Task4.V16.Test/DataServiceTest.cs
@@ -26,7 +26,13 @@ namespace Tyuiu.ZhdanovaAA.Sprint5.Task4.V16.Test
             DataService ds = new DataService();
 
             string path = @"C:\DataSprint5\InPutDataFileTask4.txt";
-            File.WriteAllText(path, Convert.ToString(2));
+
+            FileInfo fileInfo = new FileInfo(path);
+            bool fileExists = fileInfo.Exists;
+            if (!fileExists)
+            {
+                File.WriteAllText(path, Convert.ToString(2));
+            }
             double res = ds.LoadFromDataFile(path);
             double wait = 0.251;
             Assert.AreEqual(wait, res);
